requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.31.0
flask-cors==4.0.0